macling hello world
linking manually using
ld -image_base 0x8000 -pagezero_size 0x8000 -arch x86_64 embedfs.o ../core/vmling.o -lm -lSystem -framework System -lnettle -o macling
