+++ Config.mk
@@ -1,7 +1,9 @@
 LING_VER := 0.3.2
 OTP_VER := 17
+ARCH := posix
+
 ifeq ($(shell uname),Darwin)
-CROSS := 1
+CROSS := 0
 else
 CROSS := 0
 endif
