+++ core/arch/arm/console.c
