+++ core/arch/posix/include/os.h
@@ -1,0 +1,4 @@
+#if defined(__APPLE__)
+#define __INT64_C(c)              (c ## L)
+#define __UINT64_C(c)             (c ## UL)
+#endif
