+++ core/arch/posix/startup.c
@@ -1,0 +1,8 @@
+
+extern void start_ling(void);
+
+int
+main(int argc, char **argv)
+{
+	start_ling();
+}
