+++ core/config/include/machine
@@ -1,0 +1,1 @@
+../../arch/posix
