+++ core/lib/misc/libc.h
@@ -5,7 +5,11 @@
 #include <limits.h>
 
 #undef weak_alias
+#ifndef __APPLE__
 #define weak_alias(old, new) \
 	extern __typeof(old) new __attribute__((weak, alias(#old)))
+#else
+#define weak_alias(old, new)
+#endif
 
 #endif
