+++ core/preload/.gitignore
@@ -1,1 +1,1 @@
-literals.bin
+literals.c
