+++ core/scripts/premod_gen
@@ -283,13 +283,15 @@ fix_select_val_atoms_lists(Ls) ->
 
 write_literals_blob(Lings, LitBlobFile) -> %% {M,Off,Num}
 	{ok,Blob} = file:open(LitBlobFile, [raw,write]),
-	{Tab,_} = lists:foldl(fun(Ling, {Tab,Off}) ->
+	{{Tab, Iolist},_} = lists:foldl(fun(Ling, {{Tab, Iolist},Off}) ->
 		Bs = [term_to_binary(Q) || Q <- Ling#m.literals],
 		Bin = list_to_binary([<<(size(B)):32,B/binary>> || B <- Bs]),
-		ok = file:write(Blob, Bin),
 		Rec = {Ling#m.mod_name,Off,length(Bs)},
-		{[Rec|Tab],Off + size(Bin)}
-	end, {[],0}, Lings),
+		{{[Rec|Tab], [Bin, Iolist]},Off + size(Bin)}
+	end, {{[], []}, 0}, Lings),
+
+    Binary = binary_to_c_iolist("literals_blob", iolist_to_binary(Iolist)),
+	ok = file:write(Blob, Binary),
 	file:close(Blob),
 	Tab.
 
@@ -321,6 +323,22 @@ write_string_table_1(Out, <<B1:32/binary,B2/binary>>) ->
 	write_string_table_1(Out, B2);
 write_string_table_1(Out, B) ->
 	io:format(Out, "\"~s\"", [hex_str(B)]).
+
+binary_to_c_iolist(Tag, Blob) ->
+    Size = integer_to_list(size(Blob)),
+
+    P1 = ["const char ", Tag, "_start[] = {\n"],
+    Chars = lists:map(fun(C) ->
+                              case is_print(C) of
+                                  true -> [$', C, $'];
+                                  false -> [integer_to_list(C)]
+                              end
+                      end, binary_to_list(Blob)),
+    Intercalated = array_format(Chars),
+
+    P2 = ["const unsigned long ", Tag, "_size = ", Size, ";\n"],
+    P3 = ["const void *", Tag, "_end = &", Tag, "_start + ", Size, ";\n"],
+    [ P1, Intercalated, "\n};\n", P2, P3 ].
 
 hex_str(B) ->
 	lists:map(fun(C) ->
@@ -418,5 +436,13 @@ bif_type("gc_bif_" ++_, 2) -> "BIF_TYPE_GC_2";
 bif_type("gc_bif_" ++_, 3) -> "BIF_TYPE_GC_3";
 bif_type(Impl, Arity) ->
 	erlang:error({no_bif_type,Impl,Arity}).
+
+array_format(Xs) -> ["    ", intersperse(<<", ">>, <<"\n    ">>, 10, Xs)].
+intersperse(Sep, GroupSep, Count, Xs) -> nice_intersperse(Sep, GroupSep, Count, Xs, 0).
+
+nice_intersperse(_,  _, _, [],  _)  -> [];
+nice_intersperse(_,  _, _, [X], _)  -> [X];
+nice_intersperse(Sep, GroupSep, Count, [X|Xs], Count) -> [X, [Sep, GroupSep]|nice_intersperse(Sep, GroupSep, Count, Xs, 0)];
+nice_intersperse(Sep, GroupSep, Count, [X|Xs], N)     -> [X, Sep|nice_intersperse(Sep, GroupSep, Count, Xs, N + 1)].
 
 %%EOF
